Add ccdf() and make_dependent() for a Mixture of copulas
make_dependent() is a function, that assigns each pair of samples
to a single copula from the mixture, and then applies inverse CCDF
to make a pair of samples dependent.
It is a cheap way to generate samples from a mixture,
since ppcf for a mixture is not so easy to calculate
(requires iterative solver).

Minor:
1. Reorganise data_load: subdivided into data_ and model_load utils
2. Make CCDFs of Frank and Clayton 'clean' functions: avoid overwriting samples
tensor.
3. Rewrite rsample() for a mixture with a more logical indexing.
4. Add NextLayer.ipynb: a notebook that transforms the data to the next
layer.

TODO:
Fix data loader, as it still somewhere discards some samples on the
edges