[skip travis] Add heuristic model selection
Model selection first tries Frank copula (20-30 sec) to check if there
is any dependence. Then it tries a mixture of 4 corners (either Clayton
or Gumbel) + Frank copula + Independence (~ 6 min). Selects the best model,
reduces it (throws away copulas with low concentration in a mixture),
and tries to substitute each corner with the other copula type (Clayton
vs Gumbel, ~2-8 min). Finally, when the best combination of corners is selected
and if Frank copula is a part of the model, it tries to substitute Frank
with Gauss (~2-3 min). Overall, the selection process never takes more
than 20 minutes on RTX2080 (for the most complex data).

Minor: Reduce number of WAIC samples for large models to fit into GPU memory